Split the word folds into a module, and say what they are
The word-folding file never stated its organizing idea. A word list is a matrix
over GF(2), and every fold in it contracts one axis of that matrix. Nothing said
so, so each doc comment re-derived the framing, and the two axes sat interleaved
in one file with the two helpers they share.

Give the module a header that states it once, and one file per thing: the
lookup tables, the output buffer, and one file per axis. Each axis keeps the
tests that exercise it, and the reference they check against.

Both one-shot wrappers go. Each was a single call on its axis's folder, so a
caller writes that call instead. The word-axis one had no callers left at all
outside its own tests, and dropping the pair leaves the folder as the one way in,
which is the shape that stops a caller rebuilding tables it could have reused.

The chunk accumulator becomes a method on the folder that owns the tables it
reads, rather than a free function taking them.

The split broke four doc links that pointed at items now in sibling files. They
are rewritten as plain English rather than repaired, since a link to a sibling
goes stale on rename the same way prose would.
